[P6-135] refactor: formatPrice 테스트 코드 경로 변경
